--- CollatzProject/src/project/annotations/ProcessAPI.java
@@ -1,12 +1,0 @@
-package project.annotations;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import java.lang.annotation.ElementType;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface ProcessAPI {
-	
-}
