--- CollatzProject/src/project/api/conceptual/ComputeRequest.java
@@ -1,5 +1,0 @@
-package project.api.conceptual;
-
-public interface ComputeRequest {
-    int getInputNumber();
-}
