--- CollatzProject/src/project/api/network/UserComputeAPI.java
@@ -1,6 +1,0 @@
-package project.api.network;
-import project.annotations.NetworkAPI;
-@NetworkAPI
-public interface UserComputeAPI {
-    UserComputeResult processInput(UserComputeRequest request);
-}
