--- CollatzProject/src/project/api/process/DataStorageComputeAPI.java
@@ -1,8 +1,0 @@
-package project.api.process;
-
-import project.annotations.ProcessAPI;
-
-@ProcessAPI
-public interface DataStorageComputeAPI {
-    ProcessResult processData(ProcessRequest request);
-}
