--- CollatzProject/src/project/api/process/ProcessResult.java
@@ -1,19 +1,0 @@
-package project.api.process;
-
-public class ProcessResult {
-    private final boolean success;
-    private final String message;
-
-    public ProcessResult(boolean success, String message) {
-        this.success = success;
-        this.message = message;
-    }
-
-    public boolean isSuccess() {
-        return success;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
