--- CollatzProject/src/project/prototype/DataStorageComputeAPIPrototype.java
@@ -1,26 +1,0 @@
-package project.prototype;
-import project.annotations.ProcessAPIPrototype;
-import project.api.process.DataStorageComputeAPI;
-import project.api.process.ProcessRequest;
-import project.api.process.ProcessResult;
-import java.util.Arrays;
-import java.util.List;
-public class DataStorageComputeAPIPrototype {
-
-    @ProcessAPIPrototype
-    public ProcessResult prototypeProcessData(DataStorageComputeAPI api) {
-        // Mock request for demonstration purposes
-        ProcessRequest mockRequest = new ProcessRequest() {
-            public List<Integer> getInputData() {
-                return Arrays.asList(6); 
-            }
-            public String getOutputDestination() {
-                return "output_data.txt";
-            }
-        };
-
-        System.out.println("Mock processing input: " + mockRequest.getInputData());
-        System.out.println("Writing output to: " + mockRequest.getOutputDestination());
-        return new ProcessResult(true, "Prototype complete");
-    }
-}
