+++ src/project/api/conceptual/ComputeEngineAPI.java
@@ -1,0 +1,8 @@
+package project.api.conceptual;
+
+import project.annotations.ConceptualAPI;
+
+@ConceptualAPI
+public interface ComputeEngineAPI {
+    ComputeResult computeCollatz(ComputeRequest request);
+}
