+++ src/project/api/conceptual/ComputeRequest.java
@@ -1,0 +1,5 @@
+package project.api.conceptual;
+
+public interface ComputeRequest {
+    int getInputNumber();
+}
