+++ src/project/api/network/UserComputeAPI.java
@@ -1,0 +1,6 @@
+package project.api.network;
+import project.annotations.NetworkAPI;
+@NetworkAPI
+public interface UserComputeAPI {
+    UserComputeResult processInput(UserComputeRequest request);
+}
