+++ src/project/api/network/UserComputeRequest.java
@@ -1,0 +1,7 @@
+package project.api.network;
+
+public interface UserComputeRequest {
+    String getInputSource();        // input
+    String getOutputDelimiter();    // if null
+    String getOutputDestination();  // Output
+}
