+++ src/project/api/process/DataStorageComputeAPI.java
@@ -1,0 +1,8 @@
+package project.api.process;
+
+import project.annotations.ProcessAPI;
+
+@ProcessAPI
+public interface DataStorageComputeAPI {
+    ProcessResult processData(ProcessRequest request);
+}
