+++ src/project/api/process/ProcessRequest.java
@@ -1,0 +1,8 @@
+package project.api.process;
+
+import java.util.List;
+
+public interface ProcessRequest {
+    List<Integer> getInputData();        
+    String getOutputDestination();       
+}
