+++ src/src/project/annotations/ConceptualAPIPrototype.java
@@ -1,0 +1,11 @@
+package project.annotations;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import java.lang.annotation.ElementType;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface ConceptualAPIPrototype { 
+	
+}
